Refactor enum from types to constant file
