+++ libs/application/templates/health-insurance/src/types.ts
@@ -1,23 +1,6 @@
 import { FieldBaseProps, FormText } from '@island.is/application/core'
 import { NationalRegistryUser, UserProfile } from '@island.is/api/schema'
-
-export enum StatusTypes {
-  PENSIONER = 'pensioner',
-  STUDENT = 'student',
-  OTHER = 'other',
-  EMPLOYED = 'employed',
-}
-
-export enum NordicCountries {
-  NORWAY = 'Norway',
-  DENMARK = 'Denmark',
-  SWEDEN = 'Sweden',
-  FINLAND = 'Finland',
-  FAROE_ISLANDS = 'Faroe Islands',
-  GREENLAND = 'Greenland',
-  ALAND = 'Åland Islands',
-  SVALBARD = 'Svalbard and Jan Mayen', // because this is in the list of countries we get from api
-}
+import { StatusTypes } from './constants'
 
 export interface Status {
   type: StatusTypes
